Move tests from ir.rs to tests.rs in backend
