+++ yjit/src/backend/ir.rs
@@ -1027,46 +1027,3 @@ def_push_2_opnd!(csel_g, Op::CSelG);
 def_push_2_opnd!(csel_ge, Op::CSelGE);
 def_push_0_opnd_no_out!(frame_setup, Op::FrameSetup);
 def_push_0_opnd_no_out!(frame_teardown, Op::FrameTeardown);
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_draining_iterator() {
-        let mut asm = Assembler::new();
-
-        asm.load(Opnd::None);
-        asm.store(Opnd::None, Opnd::None);
-        asm.add(Opnd::None, Opnd::None);
-
-        let mut iter = asm.into_draining_iter();
-
-        while let Some((index, insn)) = iter.next_unmapped() {
-            match index {
-                0 => assert_eq!(insn.op, Op::Load),
-                1 => assert_eq!(insn.op, Op::Store),
-                2 => assert_eq!(insn.op, Op::Add),
-                _ => panic!("Unexpected instruction index"),
-            };
-        }
-    }
-
-    #[test]
-    fn test_lookback_iterator() {
-        let mut asm = Assembler::new();
-
-        asm.load(Opnd::None);
-        asm.store(Opnd::None, Opnd::None);
-        asm.store(Opnd::None, Opnd::None);
-
-        let mut iter = asm.into_lookback_iter();
-
-        while let Some((index, insn)) = iter.next_unmapped() {
-            if index > 0 {
-                assert_eq!(iter.get_previous().unwrap().opnds[0], Opnd::None);
-                assert_eq!(insn.op, Op::Store);
-            }
-        }
-    }
-}
